Gets all posts from server, and populates them in the feed
