+++ app/src/main/java/com/veephealthoutloud/healthoutloud/Interfaces/VolleyCallback.java
@@ -1,0 +1,11 @@
+package com.veephealthoutloud.healthoutloud.Interfaces;
+
+import org.json.JSONArray;
+
+/**
+ * Created by Arjun Mittal on 6/10/2018.
+ */
+
+public interface VolleyCallback {
+    void onSuccess(JSONArray result);
+}
